# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kanban API",
    "version": "v1",
    "summary": "Read and change a board from a script.",
    "description": "Hand-written, not generated. Two tests hold it to the code: one fails when a route is registered and missing here or the other way round, one validates every handler response in the test suite against the schema below.\n\nThere are no tokens. The API grants what the board grants, which is everything to whoever can reach the port, so a deployment puts the same proxy or Cloudflare Access in front of it. Where identity is configured it is used for two things: \"@me\" as an assignee, and the author of a comment.\n\nA POST or PUT body must be application/json. Unknown fields are refused with a 400 that names the field, because \"titel\" for \"title\" would otherwise store a card with no title and nobody would find out until they read the board."
  },
  "servers": [
    {
      "url": "/",
      "description": "The board itself. The paths below are absolute, so they work against any host that serves it."
    }
  ],
...
  "paths": {
...
    "/api/v1/boards/{board}/columns": {
      "parameters": [
        {
          "$ref": "#/components/parameters/board"
        }
      ],
      "post": {
        "operationId": "createColumn",
        "tags": [
          "boards"
        ],
        "summary": "Add a column at the end.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ColumnInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The new column.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Column"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/boards/{board}/columns/order": {
      "parameters": [
        {
          "$ref": "#/components/parameters/board"
        }
      ],
      "put": {
        "operationId": "reorderColumns",
        "tags": [
          "boards"
        ],
        "summary": "Set the order of the columns.",
        "description": "Left to right, and every column of the board exactly once. A list that misses one, repeats one, or names a column of another board is a 400, because a board with two columns in third place is not a board anybody can draw.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Reordered."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/boards/{board}/columns/{column}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/board"
        },
        {
          "$ref": "#/components/parameters/column"
        }
      ],
      "patch": {
        "operationId": "updateColumn",
        "tags": [
          "boards"
        ],
        "summary": "Rename a column or change its WIP limit.",
        "description": "Absent fields keep the value they have, so a rename need not repeat the limit. Lowering the limit below the number of cards already in the column is allowed; the limit is checked when a card arrives.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ColumnPatch"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Updated."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteColumn",
        "tags": [
          "boards"
        ],
        "summary": "Delete a column.",
        "description": "The cards in it are deleted with it unless move_to names another column of the same board.",
        "parameters": [
          {
            "name": "move_to",
            "in": "query",
            "required": false,
            "description": "Column id on the same board that inherits the cards.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Gone."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/boards/{board}/labels": {
      "parameters": [
        {
          "$ref": "#/components/parameters/board"
        }
      ],
      "post": {
        "operationId": "createLabel",
        "tags": [
          "boards"
        ],
        "summary": "Create a label on the board.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The new label.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Label"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/boards/{board}/labels/{label}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/board"
        },
        {
          "$ref": "#/components/parameters/label"
        }
      ],
      "patch": {
        "operationId": "updateLabel",
        "tags": [
          "boards"
        ],
        "summary": "Rename a label or recolour it.",
        "description": "Absent fields keep the value they have, so a rename need not repeat the colour.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelPatch"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Updated."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteLabel",
        "tags": [
          "boards"
        ],
        "summary": "Delete a label.",
        "description": "It comes off every card that carries it.",
        "responses": {
          "204": {
            "description": "Gone."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "board": {
        "name": "board",
        "in": "path",
        "required": true,
        "description": "The board's slug, the same one its page URL uses.",
        "schema": {
          "type": "string"
        },
        "example": "board"
      },
      "column": {
        "name": "column",
        "in": "path",
        "required": true,
        "description": "Column id. It has to be a column of this board; one from another board is a 404.",
        "schema": {
          "type": "string"
        }
      },
      "label": {
        "name": "label",
        "in": "path",
        "required": true,
        "description": "Label id. It has to belong to the board in the path, or to the card's board where the path has a card.",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "The request is malformed, a field is unknown, or the service refused a value. field names the offending input where there is one.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "No such board, column, label, card or comment. A column or label id that exists on another board is also this.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "UnsupportedMediaType": {
        "description": "Content-Type is set to something other than application/json.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error": {
        "description": "Every failure has this body, including a 404 for an endpoint that does not exist and a 500.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "What went wrong, in one sentence."
          },
          "field": {
            "type": "string",
            "description": "The input the service refused, when it named one. Absent otherwise."
          }
        }
      },
...
      "Column": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "position",
          "wip_limit",
          "stops_clock"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "description": "1 is the first column."
          },
          "wip_limit": {
            "type": "integer",
            "description": "0 for a column that holds as many cards as you like."
          },
          "stops_clock": {
            "type": "boolean",
            "description": "A column where the response-time clock does not run, such as Done."
          }
        }
      },
...
      "Label": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "color"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string",
            "description": "A hex colour, lower case. Empty for the default the page picks."
          }
        }
      },
...
      "ColumnInput": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "wip_limit": {
            "type": "integer",
            "minimum": 0,
            "description": "0, or left out, for no limit."
          },
          "stops_clock": {
            "type": "boolean",
            "description": "Left out, the clock runs in this column."
          }
        }
      },
...
      "OrderInput": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "order"
        ],
        "properties": {
          "order": {
            "type": "array",
            "description": "Ids, in the order they should end up in.",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "LabelInput": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "color": {
            "type": "string",
            "pattern": "^(?:#(?:[0-9a-fA-F]{3}|[0-9a-fA-F]{6}))?$",
            "description": "A hex colour such as #3b82f6, or empty for the default."
          }
        }
      },
...
      "ColumnPatch": {
        "type": "object",
        "additionalProperties": false,
        "description": "A field left out, or sent as null, keeps its current value.",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ],
            "minLength": 1,
            "maxLength": 100
          },
          "wip_limit": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0,
            "description": "0 for no limit."
          },
          "stops_clock": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "true takes the column out of the response-time promise."
          }
        }
      },
      "LabelPatch": {
        "type": "object",
        "additionalProperties": false,
        "description": "A field left out, or sent as null, keeps its current value.",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ],
            "minLength": 1,
            "maxLength": 100
          },
          "color": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^(?:#(?:[0-9a-fA-F]{3}|[0-9a-fA-F]{6}))?$",
            "description": "A hex colour such as #3b82f6, or empty for the default."
          }
        }
      }
    }
  }
}